backend/requirements.txt
fastapi==0.128.6
uvicorn==0.40.0
google-genai==1.62.0
python-multipart==0.0.22
python-dotenv==1.2.1